# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "gourmandise.mgueye-ba.v70208.campus-centre.fr",
  "basePath": "/api/",
  "schemes": [
    "https"
  ],
  "tags": [
    {
      "name": "Produit",
      "description": "Opérations liées aux produits"
    },
    {
      "name": "Client",
      "description": "Opérations liées aux clients"
    },
    {
      "name": "Commande",
      "description": "Opérations liées aux commandes"
    },
    {
      "name": "Authentification",
      "description": "Opérations liées à l'identification"
    }
  ],
  "paths": {
    "products": {
      "get": {
        "description": "Avoir la liste entière des produits.",
        "tags": [
          "Produit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search/products": {
      "get": {
        "description": "Recherche de produits par nom ou description.",
        "tags": [
          "Produit"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/paginated": {
      "get": {
        "description": "Obtenir une liste paginée des produits.",
        "tags": [
          "Produit"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "default": 10
          },
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/product/{ref}": {
      "get": {
        "description": "Récupérer un produit grâce à sa référence.",
        "tags": [
          "Produit"
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/count": {
      "get": {
        "description": "Récupérer le nombre total de produits.",
        "tags": [
          "Produit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/vendeurs": {
      "get": {
        "description": "Récupérer la liste des vendeurs.",
        "tags": [
          "Produit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/client": {
      "post": {
        "description": "Récupérer sa fiche client à l'aide de son code client à condition d'être connecté.",
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "codec": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/dernierescommandes": {
      "post": {
        "description": "Permet à l'utilisateur de récupérer ses 10 dernières commandes à condition d'être connecté.",
        "tags": [
          "Commande"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "codec": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "Permet de se créer un compte.",
        "tags": [
          "Authentification"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nom": {
                  "type": "string"
                },
                "adresse": {
                  "type": "string"
                },
                "cp": {
                  "type": "string"
                },
                "ville": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "motdepasse": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/passcommand": {
      "post": {
        "description": "Permet à l'utilisateur de passer une commande à condition d'être connecté.",
        "tags": [
          "Commande"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "codev": {
                  "type": "string"
                },
                "date_livraison": {
                  "type": "string"
                },
                "date_commande": {
                  "type": "string"
                },
                "total_ht": {
                  "type": "number"
                },
                "total_tva": {
                  "type": "number"
                },
                "etat": {
                  "type": "string"
                },
                "com_payee": {
                  "type": "boolean"
                },
                "com_prete": {
                  "type": "boolean"
                },
                "lignecommande": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "reference": {
                        "type": "string"
                      },
                      "quantite_demandee": {
                        "type": "integer"
                      },
                      "total_ht": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Permet à un utilisateur de se connecter.",
        "tags": [
          "Authentification"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "motdepasse": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/updateclient": {
      "post": {
        "description": "Permet à l'utilisateur de mettre à jour ses informations personnelles à condition d'être connecté.",
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nom": {
                  "type": "string"
                },
                "adresse": {
                  "type": "string"
                },
                "cp": {
                  "type": "string"
                },
                "ville": {
                  "type": "string"
                },
                "telephone": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "codec": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/updatepassword": {
      "post": {
        "description": "Permet à l'utilisateur de mettre à jour son mot de passe à condition d'être connecté.",
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "motdepasse": {
                  "type": "string"
                },
                "codec": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/deleteclient": {
      "post": {
        "description": "Permet à l'utilisateur de supprimer son compte à condition d'être connecté.",
        "tags": [
          "Client"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "codec": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lastpromo": {
      "get": {
        "description": "Récupérer les trois derniers produits en promotion.",
        "tags": [
          "Produit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lastproducts": {
      "get": {
        "description": "Récupérer les dix derniers produits ajoutés.",
        "tags": [
          "Produit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lastproducts2": {
      "get": {
        "description": "Récupérer les quatre derniers produits ajoutés.",
        "tags": [
          "Produit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/checkpromo": {
      "post": {
        "description": "Vérifier la validité d'un code promo.",
        "tags": [
          "Produit"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "codepromo": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/commandpaymenton": {
      "post": {
        "description": "Marquer une commande comme payée.",
        "tags": [
          "Commande"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "numero": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/commandready": {
      "post": {
        "description": "Marquer une commande comme prête.",
        "tags": [
          "Commande"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "numero": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {}
}